Optimized dynamic sentence builder it has better structure now, components are apart and is build ready
