+++ src/components/game/utils/shuffleArray.ts
@@ -1,3 +1,8 @@
-export default function shuffleArray<T>(array: T[]): T[] {
-  return [...array].sort(() => Math.random() - 0.5);
-} 
+export default function shuffleArray<T>(array: T[]): T[]  {
+  const newArray = [...array];
+  for (let i = newArray.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
+  }
+  return newArray;
+};
